refactor(web): 抽象出GroupHeader以减少GroupDetail的代码复杂度
顺便修改了一下GroupMsgReply的样式，增加了一条上边线
