+++ src/web/containers/main/group/GroupDetail.tsx
@@ -1,28 +1,12 @@
 import React, { useMemo } from 'react';
-import { connect, DispatchProp } from 'react-redux';
-import config from '@shared/project.config';
-import Select from 'react-select';
 import ReactTooltip from 'react-tooltip';
-import {
-  showModal,
-  hideModal,
-  showAlert,
-  showSlidePanel,
-} from '@shared/redux/actions/ui';
+import { showModal, hideModal, showAlert } from '@shared/redux/actions/ui';
 import { sendMsg, sendFile } from '@shared/redux/actions/chat';
-import {
-  changeSelectGroupActor,
-  sendGroupInvite,
-} from '@shared/redux/actions/group';
 import {
   sendDiceRequest,
   sendDiceInvite,
   sendQuickDice,
 } from '@shared/redux/actions/dice';
-import GroupInvite from './GroupInvite';
-import GroupActor from './GroupActor';
-import GroupMember from './GroupMember';
-import GroupInfo from './GroupInfo';
 import DiceRequest from '../dice/DiceRequest';
 import DiceInvite from '../dice/DiceInvite';
 import ListSelect from '../../../components/ListSelect';
@@ -31,18 +15,8 @@ import MsgSendBox from '../../../components/MsgSendBox';
 import _isNil from 'lodash/isNil';
 import _get from 'lodash/get';
 import _orderBy from 'lodash/orderBy';
-import {
-  GroupActorMsgData,
-  GroupActorType,
-} from '@src/shared/redux/types/group';
 import QuickDice from '../dice/QuickDice';
-import { TRPGState } from '@redux/types/__all__';
-import GroupRule from './GroupRule';
 import { GroupInfoContext } from '@shared/context/GroupInfoContext';
-import { UserSelector } from '@web/components/modal/UserSelector';
-import { checkIsTestUser } from '@web/utils/debug-helper';
-import { GroupChannelCreate } from './modal/GroupChannelCreate';
-import { GroupMap } from './GroupMap';
 import { MsgDataManager } from '@shared/utils/msg-helper';
 import { MsgContainerContextProvider } from '@shared/context/MsgContainerContext';
 import { GroupMsgReply } from './GroupMsgReply';
@@ -57,6 +31,7 @@ import { useTRPGDispatch } from '@shared/hooks/useTRPGSelector';
 import { useCurrentUserUUID } from '@redux/hooks/useUser';
 import { getUserInfoCache } from '@shared/utils/cache-helper';
 import { getUserName } from '@shared/utils/data-helper';
+import { GroupHeader } from './GroupHeader';
 
 export const GroupDetail: React.FC = TMemo(() => {
   const groupInfo = useSelectedGroupInfo();
@@ -71,12 +46,6 @@ export const GroupDetail: React.FC = TMemo(() => {
       selfGroupActors.find((actor) => actor.uuid === selectedGroupActorUUID),
     [selfGroupActors, selectedGroupActorUUID]
   );
-
-  const handleSelectGroupActor = (item) => {
-    if (item.value !== selectedGroupActorUUID) {
-      dispatch(changeSelectGroupActor(groupUUID, item.value));
-    }
-  };
 
   const handleSendMsg = (message: string, type: MsgType) => {
     console.log('send msg:', message, 'to', groupUUID);
@@ -95,17 +64,6 @@ export const GroupDetail: React.FC = TMemo(() => {
         data: msgDataManager.toJS(),
       })
     );
-  };
-
-  /**
-   * 发送邀请入团的邀请
-   */
-  const handleSendGroupInvite = (uuids: string[]) => {
-    for (let uuid of uuids) {
-      // TODO: 需要一个待处理的group邀请列表，防止多次提交邀请
-      dispatch(sendGroupInvite(groupUUID, uuid));
-    }
-    dispatch(hideModal());
   };
 
   // 发送文件
@@ -206,137 +164,12 @@ export const GroupDetail: React.FC = TMemo(() => {
     );
   };
 
-  const actions = useMemo(() => {
-    return [
-      ...(checkIsTestUser()
-        ? [
-            {
-              name: '创建频道',
-              icon: '&#xe61c;',
-              onClick: () => {
-                dispatch(
-                  showModal(<GroupChannelCreate groupUUID={groupUUID} />)
-                );
-              },
-            },
-          ]
-        : []),
-      {
-        name: '添加团员',
-        icon: '&#xe61d;',
-        onClick: () => {
-          dispatch(
-            showModal(<UserSelector onConfirm={handleSendGroupInvite} />)
-          );
-        },
-      },
-      {
-        name: '查看团员',
-        icon: '&#xe603;',
-        component: <GroupMember />,
-      },
-      {
-        name: '人物卡',
-        icon: '&#xe61b;',
-        component: <GroupActor />,
-      },
-      {
-        name: '游戏地图',
-        icon: '&#xe6d7;',
-        component: <GroupMap />,
-      },
-      {
-        name: '游戏规则',
-        icon: '&#xe621;',
-        component: <GroupRule />,
-      },
-      {
-        name: '团信息',
-        icon: '&#xe611;',
-        component: <GroupInfo />,
-      },
-    ];
-  }, [dispatch, groupUUID]);
-
-  /**
-   * 人物卡列表
-   */
-  const options = useMemo(() => {
-    let list = [];
-    if (selfGroupActors && selfGroupActors.length > 0) {
-      list = selfGroupActors.map((item) => ({
-        value: item.uuid,
-        label: item.name,
-      }));
-    }
-    if (selectedGroupActorUUID) {
-      list.unshift({
-        value: null,
-        label: '取消选择',
-      });
-    }
-
-    return list;
-  }, [selfGroupActors, selectedGroupActorUUID]);
-
-  const headerActions = useMemo(
-    () =>
-      actions.map((item, index) => {
-        return (
-          <button
-            key={'group-action-' + index}
-            data-tip={item.name}
-            onClick={
-              item.onClick ??
-              ((e) => {
-                e.stopPropagation();
-                dispatch(showSlidePanel(item.name, item.component));
-              })
-            }
-          >
-            <i
-              className="iconfont"
-              dangerouslySetInnerHTML={{ __html: item.icon }}
-            />
-          </button>
-        );
-      }),
-    [actions, dispatch]
-  );
-
   return (
     <GroupInfoContext.Provider value={groupInfo}>
       <MsgContainerContextProvider>
         <div className="detail">
           <ReactTooltip effect="solid" />
-          <div className="group-header">
-            <div className="avatar">
-              <img
-                src={
-                  groupInfo.avatar || config.defaultImg.getGroup(groupInfo.name)
-                }
-              />
-            </div>
-            <div className="title">
-              <div className="main-title">
-                {groupInfo.name}
-                {groupInfo.status && '(开团中...)'}
-              </div>
-              <div className="sub-title">{groupInfo.sub_name}</div>
-            </div>
-            <Select
-              name="actor-select"
-              className="group-actor-select"
-              value={selectedGroupActorUUID}
-              options={options}
-              clearable={false}
-              searchable={false}
-              placeholder="请选择身份卡"
-              noResultsText="暂无身份卡..."
-              onChange={handleSelectGroupActor}
-            />
-            <div className="actions">{headerActions}</div>
-          </div>
+          <GroupHeader />
           <MsgContainer
             className="group-content"
             converseUUID={groupUUID}
